Player upgrade improvements
UpgradeMenu script logic changed, upgrade menu now controls button interactable status based on players current stats

+++ Dungeon proj/Assets/Scripts/Game/Player/PlayerMovement.cs
@@ -136,7 +136,7 @@ public class PlayerMovement : MonoBehaviour
         return _speed;
     }
 
-    public void IncreaseSpeed(int amountToAdd)
+    public void IncreaseSpeed(float amountToAdd)
     {
         _speed += amountToAdd;
     }
